+++ testspringboot/src/main/java/ru/ds/education/testspringboot/db/repository/TrashRepository.java
@@ -19,7 +19,7 @@ public interface TrashRepository extends JpaRepository<Trash, Long> {
     @Query(value = "Insert into Trash VALUES (nextval('trash_seq'), :id_tovar, :quantity, :id_cart);", nativeQuery = true)
     void add(
             @Param("id_tovar") Long id_tovar,
-            @Param("quantity") double quantity,
+            @Param("quantity") int quantity,
             @Param("id_cart") Long id_cart
     );
